Combine ncm and ncmv3 w/ some cleanup
Modify NcmClient to handle both v3 and v2. Pass "token" and
the bearer token with api_keys argument, or provide both
api_keys and api_key. Depending on which keys are present
will determine which api version to use. Provides v2 and v3
members that can be used to expliclty specify which api to use.

A little bit of code cleanup, should preserve backward compatibility

Possibly breaking change: add support for passing in a logger and
changed the default print message to stderr.  This is to allow
scripts to use stdout to print json strings so it can be piped to other
processes like 'jq'